+++ pandas_gbq/__init__.py
@@ -1,6 +1,6 @@
 from .gbq import to_gbq, read_gbq  # noqa
 
-from ._version import get_versions
+from .version import get_versions
 versions = get_versions()
 __version__ = versions.get('closest-tag', versions['version'])
 __git_revision__ = versions['full-revisionid']
